+++ src/Pim/Component/Connector/ArrayConverter/Structured/AttributeOptionStandardConverter.php
@@ -1,46 +1,34 @@
 <?php
 
-namespace Pim\Component\Connector\ArrayConverter\Flat;
+namespace Pim\Component\Connector\ArrayConverter\Structured;
 
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
 use Pim\Component\Connector\ArrayConverter\StandardArrayConverterInterface;
 use Pim\Component\Connector\Exception\ArrayConversionException;
 
 /**
- * Convert flat format to standard format for attribute option
+ * Convert structured format to standard format for attribute option
  *
  * @author    Nicolas Dupont <nicola@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * TODO: rename!
  */
-class AttributeOptionToStandardConverter implements StandardArrayConverterInterface
+class AttributeOptionStandardConverter implements StandardArrayConverterInterface
 {
-    /** @var LocaleRepositoryInterface */
-    protected $localeRepository;
-
-    /**
-     * @param LocaleRepositoryInterface $localeRepository
-     */
-    public function __construct(LocaleRepositoryInterface $localeRepository)
-    {
-        $this->localeRepository = $localeRepository;
-    }
-
     /**
      * {@inheritdoc}
      *
-     * Converts flat csv array to standard structured array:
+     * Converts yaml array to standard structured array:
      *
      * Before:
      * {
      *     'attribute': 'maximum_print_size',
      *     'code': '210_x_1219_mm',
-     *     'sort_order': '2',
-     *     'label-de_DE': '210 x 1219 mm',
-     *     'label-en_US': '210 x 1219 mm',
-     *     'label-fr_FR': '210 x 1219 mm'
+     *     'sortOrder': 2,
+     *     'labels': {
+     *         'de_DE': '210 x 1219 mm',
+     *         'en_US': '210 x 1219 mm',
+     *         'fr_FR': '210 x 1219 mm'
+     *     }
      * }
      *
      * After:
@@ -58,23 +46,10 @@ class AttributeOptionToStandardConverter implements StandardArrayConverterInterf
     public function convert(array $item, array $options = [])
     {
         $this->validate($item);
-        $convertedItem = ['labels' => []];
-        foreach ($item as $field => $data) {
-            $isLabel = false !== strpos($field, 'label-', 0);
-            if ($isLabel) {
-                $labelTokens = explode('-', $field);
-                $labelLocale = $labelTokens[1];
-                $convertedItem['labels'][$labelLocale] = $data;
-            } else {
-                $convertedItem[$field] = $data;
-            }
-        }
-        if (!isset($convertedItem['sort_order'])) {
-            $convertedItem['sort_order'] = 1;
-        }
-        $convertedItem['sort_order'] = (int) $convertedItem['sort_order'];
+        $item['sort_order'] = $item['sortOrder'];
+        unset($item['sortOrder']);
 
-        return $convertedItem;
+        return $item;
     }
 
     /**
@@ -97,12 +72,7 @@ class AttributeOptionToStandardConverter implements StandardArrayConverterInterf
             }
         }
 
-        $authorizedFields = array_merge($requiredFields, ['sort_order']);
-        $localeCodes = $this->localeRepository->getActivatedLocaleCodes();
-        foreach ($localeCodes as $code) {
-            $authorizedFields[] = 'label-' . $code;
-        }
-
+        $authorizedFields = array_merge($requiredFields, ['sortOrder', 'labels']);
         foreach ($item as $field => $data) {
             if (!in_array($field, $authorizedFields)) {
                 throw new ArrayConversionException(
@@ -113,6 +83,12 @@ class AttributeOptionToStandardConverter implements StandardArrayConverterInterf
                     )
                 );
             }
+        }
+
+        if (isset($item['labels']) && !is_array($item['labels'])) {
+            throw new ArrayConversionException(
+                sprintf('Field "labels" must be an array, data provided is "%s"', print_r($item['labels'], true))
+            );
         }
     }
 }